+++ src/Jss.js
@@ -1,4 +1,5 @@
 /* @flow */
+import isInBrowser from 'is-in-browser'
 import StyleSheet from './StyleSheet'
 import PluginsRegistry from './PluginsRegistry'
 import rulesPlugins from './plugins/rules'
@@ -6,7 +7,8 @@ import sheets from './sheets'
 import StyleRule from './rules/StyleRule'
 import createGenerateClassNameDefault from './utils/createGenerateClassName'
 import createRule from './utils/createRule'
-import findRenderer from './utils/findRenderer'
+import DomRenderer from './renderers/DomRenderer'
+import VirtualRenderer from './renderers/VirtualRenderer'
 import type {
   Rule,
   RuleFactoryOptions,
@@ -26,9 +28,13 @@ export default class Jss {
 
   plugins = new PluginsRegistry()
 
-  options: InternalJssOptions
+  options: InternalJssOptions = {
+    createGenerateClassName: createGenerateClassNameDefault,
+    Renderer: isInBrowser ? DomRenderer : VirtualRenderer,
+    plugins: []
+  }
 
-  generateClassName: generateClassName
+  generateClassName: generateClassName = createGenerateClassNameDefault()
 
   constructor(options?: JssOptions) {
     // eslint-disable-next-line prefer-spread
@@ -37,17 +43,20 @@ export default class Jss {
   }
 
   setup(options?: JssOptions = {}): this {
-    const createGenerateClassName =
-      options.createGenerateClassName ||
-      createGenerateClassNameDefault
-    this.generateClassName = createGenerateClassName()
-    this.options = {
-      ...options,
-      createGenerateClassName,
-      Renderer: findRenderer(options)
+    if (options.createGenerateClassName) {
+      this.options.createGenerateClassName = options.createGenerateClassName
+      // $FlowFixMe
+      this.generateClassName = options.createGenerateClassName()
     }
+
+    if (options.insertionPoint != null) this.options.insertionPoint = options.insertionPoint
+    if (options.virtual || options.Renderer) {
+      this.options.Renderer = options.Renderer || (options.virtual ? VirtualRenderer : DomRenderer)
+    }
+
     // eslint-disable-next-line prefer-spread
     if (options.plugins) this.use.apply(this, options.plugins)
+
     return this
   }
 
@@ -68,6 +77,7 @@ export default class Jss {
       index
     })
     this.plugins.onProcessSheet(sheet)
+
     return sheet
   }
 
@@ -114,7 +124,14 @@ export default class Jss {
    * Register plugin. Passed function will be invoked with a rule instance.
    */
   use(...plugins: Array<Plugin>): this {
-    plugins.forEach(plugin => this.plugins.use(plugin))
+    plugins.forEach(plugin => {
+      // Avoids applying same plugin twice, at least based on ref.
+      if (this.options.plugins.indexOf(plugin) === -1) {
+        this.options.plugins.push(plugin)
+        this.plugins.use(plugin)
+      }
+    })
+
     return this
   }
 }